basic search on cities

+++ src/main/java/com/tempotalent/api/repositories/CompanyRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.tempotalent.api.models.Company;
 
-public interface CompanyRepository extends JpaRepository<Company, Long> { }
+public interface CompanyRepository extends JpaRepository<Company, Integer> { }
