+++ src/main/java/com/tempotalent/api/repositories/RecruiterRepository.java
@@ -3,4 +3,4 @@ package com.tempotalent.api.repositories;
 import com.tempotalent.api.models.Recruiter;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecruiterRepository extends JpaRepository<Recruiter, Long> {}
+public interface RecruiterRepository extends JpaRepository<Recruiter, Integer> {}
